Now the logic is in a Thread, not a KeyListener
