+++ src/com/company/Main.java
@@ -9,6 +9,7 @@ public class Main {
     public static void main(String[] args){
         System.out.println("Iniciando Juego de SSDD...");
 
-        Game juego = new Game();
+        MainThread m = new MainThread();
+        m.start();
     }
 }
